+++ src/app/forms/select-box/terra-select-box.component.ts
@@ -33,6 +33,7 @@ export class TerraSelectBoxComponent implements OnInit, OnChanges
     @Input() inputName:string;
     @Input() inputIsRequired:boolean;
     @Input() inputIsDisabled:boolean;
+    @Input() inputIsSmall:boolean;
     @Input() inputTooltipText:string;
     @Input() inputTooltipPlacement:string;
     @Input() inputListBoxValues:Array<TerraSelectBoxValueInterface>;
@@ -89,6 +90,7 @@ export class TerraSelectBoxComponent implements OnInit, OnChanges
 
         this._isInit = false;
         this.inputTooltipPlacement = 'top';
+        this.inputIsSmall = false;
     }
 
     ngOnInit()
@@ -181,7 +183,7 @@ export class TerraSelectBoxComponent implements OnInit, OnChanges
     {
         if(this._toggleOpen !== value && value == true)
         {
-            document.addEventListener('click', this.clickListener);
+            document.addEventListener('click', this.clickListener, true);
         }
         else if(this._toggleOpen !== value && value == false)
         {