+++ src/app/table/simple/row/terra-simple-table-row.interface.ts
@@ -3,8 +3,12 @@ import { TerraSimpleTableCellInterface } from '../cell/terra-simple-table-cell.i
 /**
  * @author mkunze
  */
-export interface TerraSimpleTableRowInterface
+export interface TerraSimpleTableRowInterface<D>
 {
-    cellList:Array<TerraSimpleTableCellInterface>,
+    cellList:Array<TerraSimpleTableCellInterface>;
+    selected?:boolean;
+    value?:D;
     textColorCss?:string;
+    backgroundColor?:string;
+    disabled?:boolean;
 }