+++ src/app/table/simple/terra-simple-table.component.ts
@@ -1,21 +1,39 @@
 import {
     Component,
-    Input
+    EventEmitter,
+    Input,
+    Output,
+    ViewChild
 } from '@angular/core';
 import { TerraSimpleTableHeaderCellInterface } from './cell/terra-simple-table-header-cell.interface';
 import { TerraSimpleTableRowInterface } from './row/terra-simple-table-row.interface';
+import { TerraCheckboxComponent } from '../../forms/checkbox/terra-checkbox.component';
+import { Key } from 'ts-keycode-enum';
 
 @Component({
     selector: 'terra-simple-table',
     styles:   [require('./terra-simple-table.component.scss')],
     template: require('./terra-simple-table.component.html')
 })
-export class TerraSimpleTableComponent
+export class TerraSimpleTableComponent<D>
 {
     @Input() inputHeaderList:Array<TerraSimpleTableHeaderCellInterface>;
-    @Input() inputRowList:Array<TerraSimpleTableRowInterface>;
+    @Input() inputRowList:Array<TerraSimpleTableRowInterface<D>>;
     @Input() inputUseHighlighting:boolean = false;
     @Input() inputIsStriped:boolean = false;
+    @Input() inputHasCheckboxes:boolean = false;
+    @Input() inputEnableHotkeys:boolean = false;
+    @Input() inputHighlightedRow:TerraSimpleTableRowInterface<D>;
+
+    @Output() outputHeaderCheckBoxChanged:EventEmitter<boolean> = new EventEmitter();
+    @Output() outputRowCheckBoxChanged:EventEmitter<TerraSimpleTableRowInterface<D>> = new EventEmitter();
+    @Output() outputRowClicked:EventEmitter<TerraSimpleTableRowInterface<D>> = new EventEmitter();
+    @Output() outputHighlightedRowChange:EventEmitter<TerraSimpleTableRowInterface<D>> = new EventEmitter();
+
+    @ViewChild('viewChildHeaderCheckbox') viewChildHeaderCheckbox:TerraCheckboxComponent;
+
+    private _isHeaderCheckboxChecked:boolean = false;
+    private _selectedRowList:Array<TerraSimpleTableRowInterface<D>> = [];
 
     constructor()
     {
@@ -46,12 +64,12 @@ export class TerraSimpleTableComponent
      *
      * @returns {Array<TerraSimpleTableRowInterface>}
      */
-    public get rowList():Array<TerraSimpleTableRowInterface>
+    public get rowList():Array<TerraSimpleTableRowInterface<D>>
     {
         return this.inputRowList;
     }
 
-    public set rowList(value:Array<TerraSimpleTableRowInterface>)
+    public set rowList(value:Array<TerraSimpleTableRowInterface<D>>)
     {
         this.inputRowList = value;
     }
@@ -61,7 +79,7 @@ export class TerraSimpleTableComponent
      *
      * @param rowToDelete
      */
-    public deleteRow(rowToDelete:TerraSimpleTableRowInterface):void
+    public deleteRow(rowToDelete:TerraSimpleTableRowInterface<D>):void
     {
         let index = this.inputRowList.indexOf(rowToDelete);
 
@@ -76,5 +94,136 @@ export class TerraSimpleTableComponent
         }
 
         return 'top';
+    }
+
+    private onHeaderCheckboxChange(isChecked:boolean):void
+    {
+        this._isHeaderCheckboxChecked = isChecked;
+
+        this.outputHeaderCheckBoxChanged.emit(isChecked);
+
+        this.inputRowList.forEach((row) =>
+        {
+            if(!row.disabled)
+            {
+                this.changeRowState(isChecked, row);
+            }
+        });
+    }
+
+    private onRowCheckboxChange(isChecked:boolean, row:TerraSimpleTableRowInterface<D>):void
+    {
+        this.changeRowState(isChecked, row);
+        this.outputRowCheckBoxChanged.emit(row);
+
+        if(this._selectedRowList.length === 0)
+        {
+            this._isHeaderCheckboxChecked = false;
+        }
+        else if(this._selectedRowList.length > 0 && this.inputRowList.length === this._selectedRowList.length)
+        {
+            this._isHeaderCheckboxChecked = true;
+        }
+        else
+        {
+            this.viewChildHeaderCheckbox.isIndeterminate = true;
+        }
+    }
+
+    private onRowClick(row:TerraSimpleTableRowInterface<D>):void
+    {
+        if(this.inputUseHighlighting && !row.disabled)
+        {
+            this.inputHighlightedRow = row;
+            this.outputHighlightedRowChange.emit(this.inputHighlightedRow);
+        }
+        this.outputRowClicked.emit(row);
+    }
+
+    private onKeydown(event:KeyboardEvent):void
+    {
+        if(this.inputEnableHotkeys && this.inputUseHighlighting && this.inputHighlightedRow)
+        {
+            if(event.which === Key.DownArrow || event.which === Key.UpArrow)
+            {
+                this.highlightSiblingRow(event.which === Key.DownArrow)
+            }
+
+            if((event.which === Key.Space || event.which === Key.Enter) && this.inputHasCheckboxes)
+            {
+                if(event.ctrlKey || event.metaKey)
+                {
+                    this.onHeaderCheckboxChange(!this._isHeaderCheckboxChecked);
+                }
+                else
+                {
+                    this.changeRowState(!this.inputHighlightedRow.selected, this.inputHighlightedRow);
+                }
+            }
+
+            event.preventDefault();
+        }
+    }
+
+    private highlightSiblingRow(nextSibling:boolean)
+    {
+        if(this.inputHighlightedRow)
+        {
+            let i:number = nextSibling ? 1 : -1;
+            let highlightIndex:number = this.inputRowList.indexOf(this.inputHighlightedRow) + i;
+
+            while(highlightIndex >= 0 && highlightIndex < this.inputRowList.length)
+            {
+                if(!this.inputRowList[highlightIndex].disabled)
+                {
+                    this.inputHighlightedRow = this.inputRowList[highlightIndex];
+                    this.outputHighlightedRowChange.emit(this.inputHighlightedRow);
+                    break;
+                }
+                highlightIndex += i;
+            }
+
+            /*
+            if(nextSibling && highlightIndex < this.inputRowList.length - 1)
+            {
+                this.inputHighlightedRow = this.inputRowList[highlightIndex + 1];
+                this.outputHighlightedRowChange.emit(this.inputHighlightedRow);
+            }
+            if(!nextSibling && highlightIndex > 0)
+            {
+                this.inputHighlightedRow = this.inputRowList[highlightIndex - 1];
+                this.outputHighlightedRowChange.emit(this.inputHighlightedRow);
+            }
+            */
+        }
+    }
+
+    private changeRowState(isChecked:boolean, rowToChange:TerraSimpleTableRowInterface<D>):void
+    {
+        rowToChange.selected = isChecked;
+
+        let rowFound:boolean = false;
+
+        this._selectedRowList.forEach((row) =>
+        {
+            if(row === rowToChange)
+            {
+                rowFound = true;
+            }
+        });
+
+        if(rowToChange.selected)
+        {
+            if(!rowFound)
+            {
+                this._selectedRowList.push(rowToChange);
+            }
+        }
+        else
+        {
+            let index = this._selectedRowList.indexOf(rowToChange);
+
+            this._selectedRowList.splice(index, 1);
+        }
     }
 }