+++ CHANGELOG.md
@@ -5,6 +5,15 @@
 * **terra-pager** clicks on paging buttons are now debounced
 * **terra-data-table** only one paging request can now be pending at the same time
 * **terra-button-with-options** new component, that looks like a usual button, but opens a dropdown with options on click.
+* **terra-simple-table** Add hotkeys to control simple table:
+Enable hotkeys by setting `[inputEnableHotkeys]="true"` and `[inputUseHighlighting]="true"`
+Use `arrow up` / `arrow down` to navigate throught table rows. 
+Use `space` or `enter` to toggle checkbox of selected row 
+or `ctrl`/`cmd` + `space`/`enter` to toggle checkbox in table header.
+Toggling checkboxes required `[inputHasCheckbox]="true"`
+* **terra-simple-table** Add option to disable single rows to `TerraSimpleTableRowInterface`. Disabled rows may not be highlighted (using `inputUseHighlighting`) or selected when having checkboxes activated (`inputHasCheckbox`)
+* **terra-data-table-no-result-notice-component** new component, that can be used to display a notice whenever no results are available
+
 
 ### Bug Fixes
 * **terra-multi-split-view** 
@@ -12,6 +21,7 @@
 	- prevent adding two views with the same name to the same hierarchy level
 * **terra-navigator** reduce height if search bar is enabled
 * **terra-base-service** avoid to show error message if code of error is null
+* **terra-select-box** Drop down closes if user had clicked on another select box
 
 <a name="1.4.2"></a>
 # 1.4.2 (20.09.2017)